components: unify builder vocabulary drift, deprecate old spellings (#163)
Picks one canonical name per drifted concept and deprecates the others
with #[deprecated] shims, mirroring the precedent radio already set for
.active() -> .selected():

- dialog: on_dismiss -> on_close (matches alert/notification)
- sidebar_item/button/content_button: .active(bool) -> .selected(bool)
- group_box: .outline() -> .border() (matches card)
- scroll_container: .fill(bool) -> .fill_viewport(bool), freeing .fill()
  to mean only "filled background" as group_box already uses it
- spinner/skeleton: size builders take f32, matching status_dot/icon
  and Theme::Density
- tooltip: delay_ms(u64) deprecated in favor of delay(Duration)

Every internal call site (demos, gallery example, doc comments) moved
to the canonical name; each rename has a regression test asserting the
deprecated shim still delegates correctly.

+++ src/components/skeleton/view.rs
@@ -91,10 +91,10 @@ impl Skeleton {
 
     /// Render as a circle of the given diameter in px — the idiom for avatar
     /// placeholders.
-    pub fn circle(mut self, diameter: f64) -> Self {
+    pub fn circle(mut self, diameter: f32) -> Self {
         self.shape = SkeletonShape::Circle;
-        self.width = Some(diameter);
-        self.height = Some(diameter);
+        self.width = Some(f64::from(diameter));
+        self.height = Some(f64::from(diameter));
         self
     }
 
@@ -106,28 +106,28 @@ impl Skeleton {
 
     /// Set a fixed width in px. Without this, the placeholder fills the
     /// available width.
-    pub fn width(mut self, px: f64) -> Self {
-        self.width = Some(px);
+    pub fn width(mut self, px: f32) -> Self {
+        self.width = Some(f64::from(px));
         self
     }
 
     /// Set a fixed height in px. Defaults to one line of body text.
-    pub fn height(mut self, px: f64) -> Self {
-        self.height = Some(px);
+    pub fn height(mut self, px: f32) -> Self {
+        self.height = Some(f64::from(px));
         self
     }
 
     /// Set both width and height to the same square size in px.
-    pub fn size(mut self, px: f64) -> Self {
-        self.width = Some(px);
-        self.height = Some(px);
+    pub fn size(mut self, px: f32) -> Self {
+        self.width = Some(f64::from(px));
+        self.height = Some(f64::from(px));
         self
     }
 
     /// Override the corner radius in px. Defaults to `theme.radius.small`.
     /// Ignored when the shape is [`SkeletonShape::Circle`].
-    pub fn rounded(mut self, px: f64) -> Self {
-        self.radius = Some(px);
+    pub fn rounded(mut self, px: f32) -> Self {
+        self.radius = Some(f64::from(px));
         self
     }
